Refactor into NestedBundles.solsticeVersion()
